+++ src/pages/Breweries/index.tsx
@@ -1,32 +1,19 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { BREWERIES } from "../../constants/idNames";
 import { URL_LOGIN } from "../../constants/urlRoutes";
 import { useNavigate } from "react-router";
 import BackButton from "../../components/BackButton";
-import breweriesService from "../../services/Breweries";
 import BreweryCard from "../../components/BreweryCard";
 import { useUserStore } from "../../stores/User/Context";
+import { observer } from "mobx-react";
 import "./style.less";
-import { BreweryInfo } from "../../services/Breweries/types";
 
-const Breweries: React.FC = () => {
-  const navigate = useNavigate();
+const Breweries: React.FC = observer(() => {
   const userStore = useUserStore();
-  const [breweries, setBreweries] = useState<BreweryInfo[]>([]);
-
-  useEffect(() => {
-    breweriesService.getAll().then((data) => setBreweries(data.data));
-  }, []);
+  const navigate = useNavigate();
 
   const handleCardDelete = (event: React.FormEvent<HTMLButtonElement>) => {
-    const breweriesUpdated = [...breweries];
-    const breweryIndex = breweries
-      .map((brewery) => brewery.id)
-      .indexOf(event.currentTarget.value);
-
-    breweriesUpdated.splice(breweryIndex, 1);
-
-    setBreweries(breweriesUpdated);
+    userStore?.deleteBrewery(userStore.user, event.currentTarget.value);
   };
 
   return (
@@ -39,7 +26,7 @@ const Breweries: React.FC = () => {
       </header>
       <main id={BREWERIES}>
         <section>
-          {breweries.map((brewery) => (
+          {userStore?.user.breweries.map((brewery) => (
             <BreweryCard
               key={brewery.id}
               brewery={brewery}
@@ -50,6 +37,6 @@ const Breweries: React.FC = () => {
       </main>
     </>
   );
-};
+});
 
 export default Breweries;